src/main/java/ec/edu/ups/modelo/Noticia.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package ec.edu.ups.modelo;

import java.awt.Image;
import java.io.Serializable;
import java.util.Date;

/**
 *
 * @author Diego Duchimaza
 */
public class Noticia implements Serializable{
    private static final long serialVersionUID = 1L;
    private Date fecha;
    private String titulo;
    private String descrpcion;
    private Image imagen;

    public Date getFecha() {
        return fecha;
    }

    public void setFecha(Date fecha) {
        this.fecha = fecha;
    }

    public String getTitulo() {
        return titulo;
    }

    public void setTitulo(String titulo) {
        this.titulo = titulo;
    }

    public String getDescrpcion() {
        return descrpcion;
    }

    public void setDescrpcion(String descrpcion) {
        this.descrpcion = descrpcion;
    }

    public Image getImagen() {
        return imagen;
    }

    public void setImagen(Image imagen) {
        this.imagen = imagen;
    }
    
    
    
}
